starting page created
