--- src/main/java/com/sami/backtochildhood/GameType.java
@@ -1,146 +1,0 @@
-/*
- * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
- * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JPanel.java to edit this template
- */
-package com.sami.backtochildhood;
-
-/**
- *
- * @author As-Sami
- */
-public class GameType extends javax.swing.JPanel {
-
-    /**
-     * Creates new form GameType
-     */
-    public GameType() {
-        initComponents();
-    }
-
-    /**
-     * This method is called from within the constructor to initialize the form.
-     * WARNING: Do NOT modify this code. The content of this method is always
-     * regenerated by the Form Editor.
-     */
-    @SuppressWarnings("unchecked")
-    // <editor-fold defaultstate="collapsed" desc="Generated Code">//GEN-BEGIN:initComponents
-    private void initComponents() {
-
-        jPanel2 = new javax.swing.JPanel();
-        TicTacToe = new javax.swing.JButton();
-        MakeABox = new javax.swing.JButton();
-        JoduModu = new javax.swing.JButton();
-        backButton = new javax.swing.JButton();
-
-        setBackground(new java.awt.Color(15, 83, 191));
-
-        jPanel2.setBackground(new java.awt.Color(1, 38, 97));
-
-        TicTacToe.setBackground(new java.awt.Color(0, 2, 69));
-        TicTacToe.setFont(new java.awt.Font("Dialog", 1, 18)); // NOI18N
-        TicTacToe.setText("Tic-tac-toe");
-        TicTacToe.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                TicTacToeActionPerformed(evt);
-            }
-        });
-
-        MakeABox.setBackground(new java.awt.Color(0, 2, 69));
-        MakeABox.setFont(new java.awt.Font("Dialog", 1, 18)); // NOI18N
-        MakeABox.setText("Make a box");
-        MakeABox.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                MakeABoxActionPerformed(evt);
-            }
-        });
-
-        JoduModu.setBackground(new java.awt.Color(0, 2, 69));
-        JoduModu.setFont(new java.awt.Font("Dialog", 1, 18)); // NOI18N
-        JoduModu.setText("Jodu Modhu");
-        JoduModu.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                JoduModuActionPerformed(evt);
-            }
-        });
-
-        backButton.setBackground(new java.awt.Color(0, 2, 69));
-        backButton.setText("Back");
-        backButton.addActionListener(new java.awt.event.ActionListener() {
-            public void actionPerformed(java.awt.event.ActionEvent evt) {
-                backButtonActionPerformed(evt);
-            }
-        });
-
-        javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
-        jPanel2.setLayout(jPanel2Layout);
-        jPanel2Layout.setHorizontalGroup(
-            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
-                .addContainerGap(83, Short.MAX_VALUE)
-                .addGroup(jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.TRAILING, false)
-                    .addComponent(MakeABox, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(TicTacToe, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
-                    .addComponent(JoduModu, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-                .addGap(72, 72, 72))
-            .addGroup(jPanel2Layout.createSequentialGroup()
-                .addGap(20, 20, 20)
-                .addComponent(backButton)
-                .addContainerGap(javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE))
-        );
-        jPanel2Layout.setVerticalGroup(
-            jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(javax.swing.GroupLayout.Alignment.TRAILING, jPanel2Layout.createSequentialGroup()
-                .addGap(26, 26, 26)
-                .addComponent(backButton)
-                .addGap(18, 18, 18)
-                .addComponent(MakeABox, javax.swing.GroupLayout.PREFERRED_SIZE, 48, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(31, 31, 31)
-                .addComponent(TicTacToe, javax.swing.GroupLayout.PREFERRED_SIZE, 48, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addGap(32, 32, 32)
-                .addComponent(JoduModu, javax.swing.GroupLayout.PREFERRED_SIZE, 48, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(82, Short.MAX_VALUE))
-        );
-
-        javax.swing.GroupLayout layout = new javax.swing.GroupLayout(this);
-        this.setLayout(layout);
-        layout.setHorizontalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGap(271, 271, 271)
-                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(309, Short.MAX_VALUE))
-        );
-        layout.setVerticalGroup(
-            layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addGroup(layout.createSequentialGroup()
-                .addGap(83, 83, 83)
-                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addContainerGap(109, Short.MAX_VALUE))
-        );
-    }// </editor-fold>//GEN-END:initComponents
-
-    private void TicTacToeActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_TicTacToeActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_TicTacToeActionPerformed
-
-    private void MakeABoxActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_MakeABoxActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_MakeABoxActionPerformed
-
-    private void JoduModuActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_JoduModuActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_JoduModuActionPerformed
-
-    private void backButtonActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_backButtonActionPerformed
-        // TODO add your handling code here:
-    }//GEN-LAST:event_backButtonActionPerformed
-
-
-    // Variables declaration - do not modify//GEN-BEGIN:variables
-    private javax.swing.JButton JoduModu;
-    private javax.swing.JButton MakeABox;
-    private javax.swing.JButton TicTacToe;
-    private javax.swing.JButton backButton;
-    private javax.swing.JPanel jPanel2;
-    // End of variables declaration//GEN-END:variables
-}
